+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 
 - Added specific run animation sprites.
 - Added sprite variations to convey punch charge.
+- Added plane to the Tiled tileset, which adds a standalone goal plane for the first level.
 
 ### Removed